requirements.txt
efficientnet-pytorch==0.7.1
matplotlib==3.7.2
numpy==1.25.1
pandas==2.0.3
Pillow==10.0.0
pretrainedmodels==0.7.4
pyparsing==3.0.9
scikit-learn==1.3.0
scipy==1.11.1
segmentation-models-pytorch==0.3.3
tensorboard==2.13.0
timm==0.9.2
torch==2.0.1
torchvision==0.15.2
